Set C bindings for unix_gethost* bindings in their own files
