+++ src/unix/lwt_unix_unix.h
@@ -43,503 +43,9 @@
 
 #include "unix_recv_send_utils.h"
 
-/* +-----------------------------------------------------------------+
-   | JOBs: get{pw,gr}{nam,uid}                                       |
-   +-----------------------------------------------------------------+ */
-
-#if !defined(__ANDROID__)
-
-static value alloc_passwd_entry(struct passwd *entry)
-{
-    value res;
-    value name = Val_unit, passwd = Val_unit, gecos = Val_unit;
-    value dir = Val_unit, shell = Val_unit;
-
-    Begin_roots5(name, passwd, gecos, dir, shell);
-    name = caml_copy_string(entry->pw_name);
-    passwd = caml_copy_string(entry->pw_passwd);
-#if !defined(__BEOS__)
-    gecos = caml_copy_string(entry->pw_gecos);
-#else
-    gecos = caml_copy_string("");
-#endif
-    dir = caml_copy_string(entry->pw_dir);
-    shell = caml_copy_string(entry->pw_shell);
-    res = caml_alloc_small(7, 0);
-    Field(res, 0) = name;
-    Field(res, 1) = passwd;
-    Field(res, 2) = Val_int(entry->pw_uid);
-    Field(res, 3) = Val_int(entry->pw_gid);
-    Field(res, 4) = gecos;
-    Field(res, 5) = dir;
-    Field(res, 6) = shell;
-    End_roots();
-    return res;
-}
-
-static value alloc_group_entry(struct group *entry)
-{
-    value res;
-    value name = Val_unit, pass = Val_unit, mem = Val_unit;
-
-    Begin_roots3(name, pass, mem);
-    name = caml_copy_string(entry->gr_name);
-    pass = caml_copy_string(entry->gr_passwd);
-    mem = caml_copy_string_array((const char **)entry->gr_mem);
-    res = caml_alloc_small(4, 0);
-    Field(res, 0) = name;
-    Field(res, 1) = pass;
-    Field(res, 2) = Val_int(entry->gr_gid);
-    Field(res, 3) = mem;
-    End_roots();
-    return res;
-}
-
-#define JOB_GET_ENTRY(INIT, FUNC, CONF, TYPE, ARG, ARG_DECL, FAIL_ARG) \
-    struct job_##FUNC {                                                \
-        struct lwt_unix_job job;                                       \
-        struct TYPE entry;                                             \
-        struct TYPE *ptr;                                              \
-        char *buffer;                                                  \
-        int result;                                                    \
-        ARG_DECL;                                                      \
-    };                                                                 \
-                                                                       \
-    static void worker_##FUNC(struct job_##FUNC *job)                  \
-    {                                                                  \
-        size_t buffer_size = sysconf(_SC_##CONF##_R_SIZE_MAX);         \
-        if (buffer_size == (size_t)-1) buffer_size = 16384;            \
-        job->buffer = (char *)lwt_unix_malloc(buffer_size);            \
-        job->result = FUNC##_r(job->ARG, &job->entry, job->buffer,     \
-                               buffer_size, &job->ptr);                \
-    }                                                                  \
-                                                                       \
-    static value result_##FUNC(struct job_##FUNC *job)                 \
-    {                                                                  \
-        int result = job->result;                                      \
-        if (result) {                                                  \
-            value arg = FAIL_ARG;                                      \
-            free(job->buffer);                                         \
-            lwt_unix_free_job(&job->job);                              \
-            unix_error(result, #FUNC, arg);                            \
-        } else if (job->ptr == NULL) {                                 \
-            free(job->buffer);                                         \
-            lwt_unix_free_job(&job->job);                              \
-            caml_raise_not_found();                                    \
-        } else {                                                       \
-            value entry = alloc_##TYPE##_entry(&job->entry);           \
-            free(job->buffer);                                         \
-            lwt_unix_free_job(&job->job);                              \
-            return entry;                                              \
-        }                                                              \
-    }                                                                  \
-                                                                       \
-    CAMLprim value lwt_unix_##FUNC##_job(value ARG)                    \
-    {                                                                  \
-        INIT;                                                          \
-        return lwt_unix_alloc_job(&job->job);                          \
-    }
-
-JOB_GET_ENTRY(LWT_UNIX_INIT_JOB_STRING(job, getpwnam, 0, name), getpwnam, GETPW,
-              passwd, name, char *name;
-              char data[], caml_copy_string(job->name))
-JOB_GET_ENTRY(LWT_UNIX_INIT_JOB_STRING(job, getgrnam, 0, name), getgrnam, GETGR,
-              group, name, char *name;
-              char data[], caml_copy_string(job->name))
-JOB_GET_ENTRY(LWT_UNIX_INIT_JOB(job, getpwuid, 0); job->uid = Int_val(uid),
-                                                   getpwuid, GETPW, passwd, uid,
-                                                   int uid, Nothing)
-JOB_GET_ENTRY(LWT_UNIX_INIT_JOB(job, getgrgid, 0); job->gid = Int_val(gid),
-                                                   getgrgid, GETGR, group, gid,
-                                                   int gid, Nothing)
-
-#else
-
-LWT_NOT_AVAILABLE1(unix_getpwnam_job)
-LWT_NOT_AVAILABLE1(unix_getgrnam_job)
-LWT_NOT_AVAILABLE1(unix_getpwuid_job)
-LWT_NOT_AVAILABLE1(unix_getgrgid_job)
-
-#endif
-
 /* Helper functions for not re-entrant functions */
 
-/* keep test in sync with discover.ml */
-#if !defined(HAS_GETHOSTBYADDR_R) || \
-    (HAS_GETHOSTBYADDR_R != 7 && HAS_GETHOSTBYADDR_R != 8)
-#define NON_R_GETHOSTBYADDR 1
-#endif
-
-/* keep test in sync with discover.ml */
-#if !defined(HAS_GETHOSTBYNAME_R) || \
-    (HAS_GETHOSTBYNAME_R != 5 && HAS_GETHOSTBYNAME_R != 6)
-#define NON_R_GETHOSTBYNAME 1
-#endif
-
-#if defined(NON_R_GETHOSTBYADDR) || defined(NON_R_GETHOSTBYNAME)
-static char **c_copy_addr_array(char **src, int addr_len)
-{
-    if (src == NULL) {
-        return NULL;
-    }
-    char **p = src;
-    size_t i = 0;
-    while (*p) {
-        i++;
-        p++;
-    }
-    const size_t ar_len = i;
-    p = malloc((ar_len + 1) * sizeof(char *));
-    if (p == NULL) {
-        return NULL;
-    }
-    for (i = 0; i < ar_len; ++i) {
-        p[i] = malloc(addr_len);
-        if (p[i] == NULL) {
-            size_t j;
-            for (j = 0; j < i; j++) {
-                free(p[j]);
-            }
-            free(p);
-            return NULL;
-        }
-        memcpy(p[i], src[i], addr_len);
-    }
-    p[ar_len] = NULL;
-    return p;
-}
-#endif
-#if !defined(HAVE_NETDB_REENTRANT) || defined(NON_R_GETHOSTBYADDR) || \
-    defined(NON_R_GETHOSTBYNAME)
-static char **c_copy_string_array(char **src)
-{
-    char **p = src;
-    size_t i = 0;
-    size_t len;
-    if (src == NULL) {
-        return NULL;
-    }
-    while (*p) {
-        i++;
-        p++;
-    }
-    len = i;
-    p = malloc((len + 1) * sizeof(char *));
-    if (p == NULL) {
-        return NULL;
-    }
-    for (i = 0; i < len; ++i) {
-        p[i] = strdup(src[i]);
-        if (p[i] == NULL) {
-            size_t j;
-            for (j = 0; j < i; j++) {
-                free(p[j]);
-            }
-            free(p);
-            return NULL;
-        }
-    }
-    p[len] = NULL;
-    return p;
-}
-
-static void c_free_string_array(char **src)
-{
-    if (src) {
-        char **p = src;
-        while (*p) {
-            free(*p);
-            ++p;
-        }
-        free(src);
-    }
-}
-
-static inline char *s_strdup(const char *s)
-{
-    return (strdup(s == NULL ? "" : s));
-}
-#endif
-
-/* +-----------------------------------------------------------------+
-   | JOB: gethostname                                                |
-   +-----------------------------------------------------------------+ */
-
-struct job_gethostname {
-    struct lwt_unix_job job;
-    char *buffer;
-    int result;
-    int error_code;
-};
-
-static void worker_gethostname(struct job_gethostname *job)
-{
-    int buffer_size = 64;
-    int err;
-
-    for (;;) {
-        job->buffer = lwt_unix_malloc(buffer_size + 1);
-
-        err = gethostname(job->buffer, buffer_size);
-
-        if (err == -1 && errno == ENAMETOOLONG) {
-            free(job->buffer);
-            buffer_size *= 2;
-        } else if (err == -1) {
-            free(job->buffer);
-            job->result = -1;
-            job->error_code = errno;
-            return;
-        } else {
-            job->buffer[buffer_size] = 0;
-            job->result = 0;
-            return;
-        }
-    }
-}
-
-static value result_gethostname(struct job_gethostname *job)
-{
-    LWT_UNIX_CHECK_JOB(job, job->result < 0, "gethostname");
-    value result = caml_copy_string(job->buffer);
-    free(job->buffer);
-    lwt_unix_free_job(&job->job);
-    return result;
-}
-
-CAMLprim value lwt_unix_gethostname_job(value Unit)
-{
-    LWT_UNIX_INIT_JOB(job, gethostname, 0);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: gethostbyname                                              |
-   +-----------------------------------------------------------------+ */
-
-#define NETDB_BUFFER_SIZE 10000
-
-struct job_gethostbyname {
-    struct lwt_unix_job job;
-    struct hostent entry;
-    struct hostent *ptr;
-#ifndef NON_R_GETHOSTBYNAME
-    char buffer[NETDB_BUFFER_SIZE];
-#endif
-    char *name;
-    char data[];
-};
-
-CAMLexport value alloc_inet_addr(struct in_addr *inaddr);
-CAMLexport value alloc_inet6_addr(struct in6_addr *inaddr);
-
-static value alloc_one_addr(char const *a)
-{
-    struct in_addr addr;
-    memmove(&addr, a, 4);
-    return alloc_inet_addr(&addr);
-}
-
-static value alloc_one_addr6(char const *a)
-{
-    struct in6_addr addr;
-    memmove(&addr, a, 16);
-    return alloc_inet6_addr(&addr);
-}
-
-static value alloc_host_entry(struct hostent *entry)
-{
-    value res;
-    value name = Val_unit, aliases = Val_unit;
-    value addr_list = Val_unit, adr = Val_unit;
-
-    Begin_roots4(name, aliases, addr_list, adr);
-    name = caml_copy_string((char *)(entry->h_name));
-    /* PR#4043: protect against buggy implementations of gethostbynamee()
-       that return a NULL pointer in h_aliases */
-    if (entry->h_aliases)
-        aliases = caml_copy_string_array((const char **)entry->h_aliases);
-    else
-        aliases = Atom(0);
-    if (entry->h_length == 16)
-        addr_list = caml_alloc_array(alloc_one_addr6,
-                                     (const char **)entry->h_addr_list);
-    else
-        addr_list =
-            caml_alloc_array(alloc_one_addr, (const char **)entry->h_addr_list);
-    res = caml_alloc_small(4, 0);
-    Field(res, 0) = name;
-    Field(res, 1) = aliases;
-    switch (entry->h_addrtype) {
-        case PF_UNIX:
-            Field(res, 2) = Val_int(0);
-            break;
-        case PF_INET:
-            Field(res, 2) = Val_int(1);
-            break;
-        default: /*PF_INET6 */
-            Field(res, 2) = Val_int(2);
-            break;
-    }
-    Field(res, 3) = addr_list;
-    End_roots();
-    return res;
-}
-
-#if defined(NON_R_GETHOSTBYADDR) || defined(NON_R_GETHOSTBYNAME)
-static struct hostent *hostent_dup(struct hostent *orig)
-{
-    if (orig == NULL) {
-        return NULL;
-    }
-    struct hostent *h = malloc(sizeof *h);
-    if (h == NULL) {
-        return NULL;
-    }
-    h->h_name = s_strdup(orig->h_name);
-    if (!h->h_name) {
-        goto nomem1;
-    }
-    if (!orig->h_aliases) {
-        h->h_aliases = NULL;
-    } else {
-        h->h_aliases = c_copy_string_array(orig->h_aliases);
-        if (!h->h_aliases) {
-            goto nomem2;
-        }
-    }
-    if (!orig->h_addr_list) {
-        h->h_addr_list = NULL;
-    } else {
-        h->h_addr_list = c_copy_addr_array(orig->h_addr_list, orig->h_length);
-        if (!h->h_addr_list) {
-            goto nomem3;
-        }
-    }
-    h->h_addrtype = orig->h_addrtype;
-    h->h_length = orig->h_length;
-    return h;
-nomem3:
-    c_free_string_array(h->h_aliases);
-nomem2:
-    free((char *)h->h_name);
-nomem1:
-    free(h);
-    return NULL;
-}
-
-static void hostent_free(struct hostent *h)
-{
-    if (h) {
-        c_free_string_array(h->h_addr_list);
-        c_free_string_array(h->h_aliases);
-        free((char *)h->h_name);
-        free(h);
-    }
-}
-#endif
-
-static void worker_gethostbyname(struct job_gethostbyname *job)
-{
-#if HAS_GETHOSTBYNAME_R == 5
-    int h_errno;
-    job->ptr = gethostbyname_r(job->name, &job->entry, job->buffer,
-                               NETDB_BUFFER_SIZE, &h_errno);
-#elif HAS_GETHOSTBYNAME_R == 6
-    int h_errno;
-    if (gethostbyname_r(job->name, &job->entry, job->buffer, NETDB_BUFFER_SIZE,
-                        &(job->ptr), &h_errno) != 0)
-        job->ptr = NULL;
-#else
-    job->ptr = gethostbyname(job->name);
-    if (job->ptr) {
-        job->ptr = hostent_dup(job->ptr);
-        if (job->ptr) {
-            job->entry = *job->ptr;
-        }
-    }
-#endif
-}
-
-static value result_gethostbyname(struct job_gethostbyname *job)
-{
-    if (job->ptr == NULL) {
-        lwt_unix_free_job(&job->job);
-        caml_raise_not_found();
-    } else {
-        value entry = alloc_host_entry(&job->entry);
-#ifdef NON_R_GETHOSTBYNAME
-        hostent_free(job->ptr);
-#endif
-        lwt_unix_free_job(&job->job);
-        return entry;
-    }
-}
-
-CAMLprim value lwt_unix_gethostbyname_job(value name)
-{
-    LWT_UNIX_INIT_JOB_STRING(job, gethostbyname, 0, name);
-    return lwt_unix_alloc_job(&(job->job));
-}
-
-/* +-----------------------------------------------------------------+
-   | JOB: gethostbyaddr                                              |
-   +-----------------------------------------------------------------+ */
-
-struct job_gethostbyaddr {
-    struct lwt_unix_job job;
-    struct in_addr addr;
-    struct hostent entry;
-    struct hostent *ptr;
-#ifndef NON_R_GETHOSTBYADDR
-    char buffer[NETDB_BUFFER_SIZE];
-#endif
-};
-
-static void worker_gethostbyaddr(struct job_gethostbyaddr *job)
-{
-#if HAS_GETHOSTBYADDR_R == 7
-    int h_errno;
-    job->ptr = gethostbyaddr_r(&job->addr, 4, AF_INET, &job->entry, job->buffer,
-                               NETDB_BUFFER_SIZE, &h_errno);
-#elif HAS_GETHOSTBYADDR_R == 8
-    int h_errno;
-    if (gethostbyaddr_r(&job->addr, 4, AF_INET, &job->entry, job->buffer,
-                        NETDB_BUFFER_SIZE, &job->ptr, &h_errno) != 0)
-        job->ptr = NULL;
-#else
-    job->ptr = gethostbyaddr(&job->addr, 4, AF_INET);
-    if (job->ptr) {
-        job->ptr = hostent_dup(job->ptr);
-        if (job->ptr) {
-            job->entry = *job->ptr;
-        }
-    }
-#endif
-}
-
-static value result_gethostbyaddr(struct job_gethostbyaddr *job)
-{
-    if (job->ptr == NULL) {
-        lwt_unix_free_job(&job->job);
-        caml_raise_not_found();
-    } else {
-        value entry = alloc_host_entry(&job->entry);
-#ifdef NON_R_GETHOSTBYADDR
-        hostent_free(job->ptr);
-#endif
-        lwt_unix_free_job(&job->job);
-        return entry;
-    }
-}
-
-CAMLprim value lwt_unix_gethostbyaddr_job(value val_addr)
-{
-    LWT_UNIX_INIT_JOB(job, gethostbyaddr, 0);
-    job->addr = GET_INET_ADDR(val_addr);
-    return lwt_unix_alloc_job(&job->job);
-}
-
+#include "unix_get_network_information_utils.h"
 /* +-----------------------------------------------------------------+
    | JOBs: getprotoby{name,number}, getservby{name,port}             |
    +-----------------------------------------------------------------+ */
