+++ src/unix/lwt_unix_windows.h
@@ -120,9 +120,6 @@ CAMLprim value lwt_unix_system_job(value cmdline)
    | Unavailable primitives                                          |
    +-----------------------------------------------------------------+ */
 
-LWT_NOT_AVAILABLE1(unix_gethostname_job)
-LWT_NOT_AVAILABLE1(unix_gethostbyname_job)
-LWT_NOT_AVAILABLE1(unix_gethostbyaddr_job)
 LWT_NOT_AVAILABLE1(unix_getprotobyname_job)
 LWT_NOT_AVAILABLE1(unix_getprotobynumber_job)
 LWT_NOT_AVAILABLE2(unix_getservbyname_job)
@@ -131,9 +128,5 @@ LWT_NOT_AVAILABLE3(unix_getaddrinfo_job)
 LWT_NOT_AVAILABLE2(unix_getnameinfo_job)
 LWT_NOT_AVAILABLE1(unix_tcgetattr_job)
 LWT_NOT_AVAILABLE3(unix_tcsetattr_job)
-LWT_NOT_AVAILABLE1(unix_getpwnam_job)
-LWT_NOT_AVAILABLE1(unix_getgrnam_job)
-LWT_NOT_AVAILABLE1(unix_getpwuid_job)
-LWT_NOT_AVAILABLE1(unix_getgrgid_job)
 LWT_NOT_AVAILABLE2(unix_bind_job)
 LWT_NOT_AVAILABLE1(unix_getcwd_job)
