+++ src/unix/windows_c/not_available.c
@@ -77,4 +77,11 @@ LWT_NOT_AVAILABLE2(unix_readdir_n_job)
 LWT_NOT_AVAILABLE1(unix_readlink_job)
 LWT_NOT_AVAILABLE3(unix_lockf_job)
 LWT_NOT_AVAILABLE1(unix_getlogin_job)
+LWT_NOT_AVAILABLE1(unix_getpwnam_job)
+LWT_NOT_AVAILABLE1(unix_getgrnam_job)
+LWT_NOT_AVAILABLE1(unix_getpwuid_job)
+LWT_NOT_AVAILABLE1(unix_getgrgid_job)
+LWT_NOT_AVAILABLE1(unix_gethostname_job)
+LWT_NOT_AVAILABLE1(unix_gethostbyname_job)
+LWT_NOT_AVAILABLE1(unix_gethostbyaddr_job)
 #endif
